(svn r16297) -Codechange: silence more ICC warnings

+++ src/dock_gui.cpp
@@ -46,7 +46,7 @@ void CcBuildCanal(bool success, TileIndex tile, uint32 p1, uint32 p2)
 
 static void PlaceDocks_Dock(TileIndex tile)
 {
-	uint32 p2 = INVALID_STATION << 16; // no station to join
+	uint32 p2 = (uint32)INVALID_STATION << 16; // no station to join
 
 	/* tile is always the land tile, so need to evaluate _thd.pos */
 	CommandContainer cmdcont = { tile, _ctrl_pressed, p2, CMD_BUILD_DOCK | CMD_MSG(STR_ERROR_CAN_T_BUILD_DOCK_HERE), CcBuildDocks, "" };
